Replace Bellman with Zexe (#6)
* feat(utils): port utils to use latest rand and Zexe

* feat(params): use Zexe and add utility methods

* feat(keypair): Refactor serialization logic and use PairingEngine

* feat(batched-accumulator): Use Zexe APIs

* add tests for calculate_position, and read/write chunks
* remove write_all in favor of using write_points_chunk explicitly in write_chunk
* return error instead of panicking in calculate_position
* remove mut requirements from some functions

* feat(cli): make phase1 work with multiple curves

* feat(cli): add support for SW6 and fix params bug

Previously we'd naively generate the G1/G2 sizes by counting the size of the ToBytes representation of each element. This is wrong, as this would give us the 'max' size of the field. Counting the bytes in a serialized buffer produced by the CanonicalSerialize trait gives the desired sizes for the compressed elements. Uncompressed elements are just twice that size.

* fix(accumulator): add missing check on pairing chunk intersections

* fix(keypair): use CanonicalSerialize::serialize to calculate hash