Refactor to match on participant in `CoordinatorState::drop_participant()`
Previously this was using `if` checks for each type of Participant,
which has the potential to break and result in a runtime error
if a new type of Participant is added. With the new code, this
will result in a compile time error, and eliminates a potential
runtime error branch.